# infrastructure/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Emerion Dashboard API
  description: >
    ERP API responsible for (1) receiving already-mapped data from
    emerion-load-service (which extracts from the legacy Firebird database)
    and persisting it into PostgreSQL, and (2) serving that data to the
    React frontend. This file is the single source of truth for the API
    contract (contract-first). Kotlin interfaces and models are generated
    from this spec via the openapi-generator-gradle-plugin
    (see build.gradle.kts `openApiGenerate` task) — do not hand-write
    controller interfaces or request/response models, edit this file instead
    and regenerate.
  version: "1.0.0"
servers:
  - url: /api/v1
...
paths:
...
  /customers:
    get:
      tags:
        - customers
      operationId: listCustomers
      summary: List customers (paginated) for the React dashboard
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: bloqueado
          in: query
          description: Filter by blocked/unblocked status
          required: false
          schema:
            type: boolean
        - name: nomeFantasia
          in: query
          description: Partial, case-insensitive match on customer nomeFantasia
          required: false
          schema:
            type: string
        - name: cnpjEmpresa
          in: query
          description: Exact match on the owning retailer's cnpjEmpresa
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPage'
  /customers/{id}:
    get:
      tags:
        - customers
      operationId: getCustomerById
      summary: Get a single customer by internal (PostgreSQL) id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/by-external-id/{externalId}:
    get:
      tags:
        - customers
      operationId: getCustomerByExternalId
      summary: Get a single customer by externalId (the legacy Firebird ID)
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /products:
    get:
      tags:
        - products
      operationId: listProducts
      summary: List products (paginated) for the React dashboard
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: nome
          in: query
          description: Partial, case-insensitive match on product nome
          required: false
          schema:
            type: string
        - name: cnpjEmpresa
          in: query
          description: Exact match on the owning retailer's cnpjEmpresa
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
  /products/{id}:
    get:
      tags:
        - products
      operationId: getProductById
      summary: Get a single product by internal (PostgreSQL) id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/by-external-id/{externalId}:
    get:
      tags:
        - products
      operationId: getProductByExternalId
      summary: Get a single product by externalId (the legacy Firebird ID)
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        externalId:
          type: string
        cnpjEmpresa:
          type: string
        nomeFantasia:
          type: string
        razaoSocial:
          type: string
        cpfCnpj:
          type: string
        inscricaoEstadual:
          type: string
          nullable: true
        regimeTributario:
          type: string
          nullable: true
        bloqueado:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    PaginationInfo:
      type: object
      properties:
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
              nullable: true
        timestamp:
          type: string
          format: date-time
...
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        externalId:
          type: string
        cnpjEmpresa:
          type: string
        nome:
          type: string
        preco:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...